crypto-bot/requirements.txt
python-binance
python-dotenv
mysql-connector-python
pandas
scikit-learn
lightgbm
xgboost
catboost
tensorflow
textblob
requests

